+++ vkconfig_cmd/main_loader.cpp
@@ -28,9 +28,9 @@ static int RunLoaderList(Configurator& configurator, const CommandLine& command_
     const auto& configurations = configurator.configurations.available_configurations;
 
     if (configurator.configurations.available_configurations.empty()) {
-        printf("vkconfig: No Vulkan Loader configuration found\n");
+        printf("vkconfig: No Vulkan Layers configuration found\n");
     } else {
-        printf("vkconfig: Vulkan Loader Configurations:\n");
+        printf("vkconfig: Vulkan Layers Configurations:\n");
         for (std::size_t i = 0, n = configurations.size(); i < n; ++i) {
             printf(" - [%zd] %s\n", i, configurations[i].key.c_str());
         }
@@ -46,7 +46,7 @@ static int GetConfigurationName(Configurator& configurator, const CommandLine& c
         const bool load_result = configurator.configurations.ImportConfiguration(
             configurator.layers, command_line.layers_configuration_path, configuration_name);
         if (!load_result) {
-            fprintf(stderr, "vkconfig: Failed to load `%s` loader configuration file...\n\n",
+            fprintf(stderr, "vkconfig: Failed to load `%s` Layers configuration file...\n\n",
                     command_line.layers_configuration_path.AbsolutePath().c_str());
 
             ::RunLoaderList(configurator, command_line);
@@ -58,7 +58,7 @@ static int GetConfigurationName(Configurator& configurator, const CommandLine& c
         if (is_index) {
             std::size_t index = std::strtoul(command_line.layers_configuration_name.c_str(), nullptr, 10);
             if (index >= configurator.configurations.available_configurations.size()) {
-                fprintf(stderr, "vkconfig: Invalid `%zd` loader configuration index...\n\n", index);
+                fprintf(stderr, "vkconfig: Invalid `%zd` Layers configuration index...\n\n", index);
 
                 ::RunLoaderList(configurator, command_line);
                 return -1;
@@ -71,7 +71,7 @@ static int GetConfigurationName(Configurator& configurator, const CommandLine& c
     }
 
     if (configurator.configurations.FindConfiguration(configuration_name) == nullptr) {
-        fprintf(stderr, "vkconfig: `%s` loader configuration not found...\n\n", configuration_name.c_str());
+        fprintf(stderr, "vkconfig: `%s` Layers configuration not found...\n\n", configuration_name.c_str());
 
         ::RunLoaderList(configurator, command_line);
         return -1;
@@ -90,12 +90,12 @@ static int RunLoaderOverride(Configurator& configurator, const CommandLine& comm
     configurator.SetActiveConfigurationName(configuration_name);
     const bool override_result = configurator.Override(OVERRIDE_AREA_ALL);
     if (override_result) {
-        printf("vkconfig: \"%s\" loader configuration applied to all Vulkan Applications.\n", configuration_name.c_str());
+        printf("vkconfig: \"%s\" Layers configuration applied to all Vulkan Applications.\n", configuration_name.c_str());
         const Configuration* configuration = configurator.GetActiveConfiguration();
 
-        if (configuration->override_layers) {
+        {
             printf("\n");
-            printf(" Including Vulkan layers:\n");
+            printf(" Vulkan layers:\n");
             printf(" (Execute Closer to the Vulkan Application)\n");
             for (std::size_t i = 0, n = configuration->parameters.size(); i < n; ++i) {
                 const Parameter& parameter = configuration->parameters[i];
@@ -104,15 +104,15 @@ static int RunLoaderOverride(Configurator& configurator, const CommandLine& comm
             printf(" (Execute Closer to the Vulkan Driver)\n");
         }
 
-        if (configuration->override_loader) {
+        if (configurator.loader_log_enabled) {
             printf("\n");
-            if (configuration->loader_log_messages_flags == 0) {
+            if (configurator.loader_log_messages_flags == 0) {
                 printf(" Including Vulkan loader messages: None\n");
             } else {
                 printf(" Including Vulkan loader messages:\n");
                 for (int i = 0, n = LOG_COUNT; i < n; ++i) {
                     LogBit bit = static_cast<LogBit>(1 << i);
-                    if (configuration->loader_log_messages_flags & bit) {
+                    if (configurator.loader_log_messages_flags & bit) {
                         printf("- %s\n", ::GetToken(bit));
                     }
                 }