+++ vkconfig_gui/tab_drivers.cpp
@@ -41,8 +41,15 @@ TabDrivers::TabDrivers(MainWindow &window, std::shared_ptr<Ui::MainWindow> ui) :
         this->ui->tab_widget->setTabToolTip(TAB_DRIVERS, "Additional Vulkan Drivers require Vulkan Loader 1.4.322 or newer");
     }
 
+    this->connect(this->ui->driver_group_box_override, SIGNAL(toggled(bool)), this, SLOT(on_driver_override_toggled(bool)));
+    this->connect(this->ui->driver_mode, SIGNAL(currentIndexChanged(int)), this, SLOT(on_driver_mode_changed(int)));
+    this->connect(this->ui->driver_forced_name, SIGNAL(currentIndexChanged(int)), this, SLOT(on_driver_name_changed(int)));
+
+    this->connect(this->ui->driver_group_box_paths, SIGNAL(toggled(bool)), this, SLOT(on_driver_paths_toggled(bool)));
     this->connect(this->ui->driver_browse_button, SIGNAL(clicked()), this, SLOT(on_driver_browse_pressed()));
     this->connect(this->ui->driver_path_lineedit, SIGNAL(returnPressed()), this, SLOT(on_driver_append_pressed()));
+
+    this->UpdateUI(UPDATE_REBUILD_UI);
 }
 
 TabDrivers::~TabDrivers() {}
@@ -50,13 +57,55 @@ TabDrivers::~TabDrivers() {}
 void TabDrivers::UpdateUI(UpdateUIMode ui_update_mode) {
     Configurator &configurator = Configurator::Get();
 
-    this->ui->driver_path_lineedit->setText(configurator.last_driver_path.RelativePath().c_str());
-    this->ui->driver_paths_list->blockSignals(true);
-
     switch (ui_update_mode) {
         case UPDATE_REFRESH_UI:
         case UPDATE_REBUILD_UI: {
+            this->ui->driver_group_box_override->blockSignals(true);
+            this->ui->driver_group_box_override->setChecked(configurator.driver_override_enabled);
+            this->ui->driver_group_box_override->blockSignals(false);
+
+            switch (configurator.driver_override_mode) {
+                default:
+                case DRIVER_MODE_SINGLE: {
+                    this->ui->driver_forced_name->blockSignals(true);
+                    this->ui->driver_forced_name->clear();
+                    for (std::size_t i = 0, n = configurator.vulkan_system_info.physicalDevices.size(); i < n; ++i) {
+                        this->ui->driver_forced_name->addItem(
+                            configurator.vulkan_system_info.physicalDevices[i].deviceName.c_str());
+                    }
+                    this->ui->driver_forced_name->setCurrentIndex(configurator.GetActiveDeviceIndex());
+                    this->ui->driver_forced_name->blockSignals(false);
+
+                    this->ui->driver_name_label->setVisible(true);
+                    this->ui->driver_forced_name->setVisible(true);
+                    this->ui->drivers_label_first->setVisible(false);
+                    this->ui->drivers_device_list->setVisible(false);
+                    this->ui->drivers_label_last->setVisible(false);
+                } break;
+                case DRIVER_MODE_SORTED: {
+                    this->ui->drivers_device_list->blockSignals(true);
+                    this->ui->drivers_device_list->clear();
+                    for (std::size_t i = 0, n = configurator.vulkan_system_info.physicalDevices.size(); i < n; ++i) {
+                        this->ui->drivers_device_list->addItem(
+                            configurator.vulkan_system_info.physicalDevices[i].deviceName.c_str());
+                    }
+                    this->ui->drivers_device_list->blockSignals(false);
+
+                    this->ui->driver_name_label->setVisible(false);
+                    this->ui->driver_forced_name->setVisible(false);
+                    this->ui->drivers_label_first->setVisible(true);
+                    this->ui->drivers_device_list->setVisible(true);
+                    this->ui->drivers_label_last->setVisible(true);
+                } break;
+            }
+
+            this->ui->driver_group_box_paths->blockSignals(true);
+            this->ui->driver_group_box_paths->setChecked(configurator.driver_paths_enabled);
+            this->ui->driver_group_box_paths->blockSignals(false);
+
             this->ui->driver_paths_list->clear();
+            this->ui->driver_path_lineedit->setText(configurator.last_driver_path.RelativePath().c_str());
+            this->ui->driver_paths_list->blockSignals(true);
 
             for (auto it = configurator.driver_paths.begin(); it != configurator.driver_paths.end(); ++it) {
                 QListWidgetItem *item_state = new QListWidgetItem;
@@ -69,10 +118,10 @@ void TabDrivers::UpdateUI(UpdateUIMode ui_update_mode) {
                 ui->driver_paths_list->addItem(item_state);
                 ui->driver_paths_list->setItemWidget(item_state, drivers_path_widget);
             }
+
+            this->ui->driver_paths_list->blockSignals(false);
         } break;
     }
-
-    this->ui->driver_paths_list->blockSignals(false);
 }
 
 void TabDrivers::CleanUI() {}
@@ -84,12 +133,51 @@ bool TabDrivers::EventFilter(QObject *target, QEvent *event) {
     return false;
 }
 
-void TabDrivers::on_paths_changed() {
-    Configurator::Get().Override(OVERRIDE_AREA_LOADER_SETTINGS_BIT);
+void TabDrivers::on_driver_override_toggled(bool checked) {
+    Configurator &configurator = Configurator::Get();
+
+    configurator.driver_override_enabled = checked;
+    configurator.Override(OVERRIDE_AREA_LOADER_SETTINGS_BIT);
+}
+
+void TabDrivers::on_driver_mode_changed(int index) {
+    Configurator &configurator = Configurator::Get();
+    configurator.driver_override_mode = static_cast<DriverMode>(index);
+    configurator.Override(OVERRIDE_AREA_LOADER_SETTINGS_BIT);
+
     this->UpdateUI(UPDATE_REBUILD_UI);
 }
 
-void TabDrivers::on_paths_toggled() { Configurator::Get().Override(OVERRIDE_AREA_LOADER_SETTINGS_BIT); }
+void TabDrivers::on_driver_name_changed(int index) {
+    Configurator &configurator = Configurator::Get();
+    configurator.driver_override_name = this->ui->driver_forced_name->itemText(index).toStdString();
+    configurator.Override(OVERRIDE_AREA_LOADER_SETTINGS_BIT);
+
+    this->UpdateUI(UPDATE_REBUILD_UI);
+}
+
+void TabDrivers::on_driver_paths_toggled(bool checked) {
+    Configurator &configurator = Configurator::Get();
+
+    configurator.driver_paths_enabled = checked;
+    configurator.Override(OVERRIDE_AREA_LOADER_SETTINGS_BIT);
+}
+
+void TabDrivers::on_paths_changed() {
+    Configurator &configurator = Configurator::Get();
+
+    configurator.UpdateVulkanSystemInfo();
+
+    this->UpdateUI(UPDATE_REBUILD_UI);
+}
+
+void TabDrivers::on_paths_toggled() {
+    Configurator &configurator = Configurator::Get();
+
+    configurator.UpdateVulkanSystemInfo();
+
+    this->UpdateUI(UPDATE_REBUILD_UI);
+}
 
 void TabDrivers::on_driver_append_pressed() {
     Configurator &configurator = Configurator::Get();
@@ -101,6 +189,7 @@ void TabDrivers::on_driver_append_pressed() {
 
     configurator.driver_paths.insert(std::pair(selected_path, true));
     configurator.last_driver_path = selected_path;
+
     configurator.UpdateVulkanSystemInfo();
 
     this->UpdateUI(UPDATE_REBUILD_UI);
@@ -119,6 +208,7 @@ void TabDrivers::on_driver_browse_pressed() {
 
     configurator.driver_paths.insert(std::pair(selected_path, true));
     configurator.last_driver_path = selected_path;
+
     configurator.UpdateVulkanSystemInfo();
 
     this->UpdateUI(UPDATE_REBUILD_UI);